+++ goldens/size-tracking/aio-payloads.json
@@ -16,7 +16,7 @@
       "uncompressed": {
         "runtime": 4343,
         "main": 450179,
-        "polyfills": 37823,
+        "polyfills": 33349,
         "styles": 70416,
         "light-theme": 77582,
         "dark-theme": 77711
